+++ .gitignore
@@ -121,6 +121,7 @@ out
 
 # Stores VSCode versions used for testing VSCode extensions
 .vscode-test
+.storage
 
 # yarn v2
 .yarn/cache
