+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "v-proxy",
-  "version": "0.0.5",
+  "version": "0.0.6",
   "description": "简单易用的跨端抓包工具",
   "main": "lib/index.js",
   "bin": {
@@ -10,13 +10,14 @@
   "scripts": {
     "start": "clear && nodemon --transpileOnly src/index.ts",
     "dev": "ts-node --transpileOnly src/index.ts",
+    "perod": "tsc && node lib/index.js",
     "build:compile": "tsc",
     "build:web": "cd src/web/ && npm run build && cd ../../",
     "copy:public": "shx mkdir -p ./lib/public && shx cp -r ./src/public/* ./lib/public",
     "copy:web": "shx mkdir -p ./lib/web/dist && shx cp -r ./src/web/dist/* ./lib/web/dist",
     "build:copy": "npm run copy:public && npm run copy:web",
     "build": "shx rm -rf lib && npm run build:compile && npm run build:copy",
-    "publish": "npm run build && npm publish",
+    "pub": "npm run build && npm publish",
     "create_crt": "ts-node --transpileOnly src/create_root_cert.ts"
   },
   "repository": {
@@ -50,6 +51,8 @@
     "@types/node-forge": "^1.0.4",
     "chalk": "^4.1.2",
     "co": "^4.6.0",
+    "daemon": "^1.1.0",
+    "daemonize-process": "^3.0.0",
     "eslint": "^8.21.0",
     "ip": "^1.1.8",
     "koa": "^2.13.4",
@@ -63,7 +66,9 @@
     "mkdirp": "^1.0.4",
     "needle": "^3.2.0",
     "node-forge": "^1.3.1",
+    "node-storage": "^0.0.9",
     "nodemon": "^2.0.19",
+    "pfork": "^0.6.0",
     "shx": "^0.3.4",
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4",
